% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade.R
\name{check_eth.to.tok_tok.fix}
\alias{check_eth.to.tok_tok.fix}
\title{Swap ETH for a Token, Given Token Qty check how much ETH you need}
\usage{
check_eth.to.tok_tok.fix(t_a, t_d, t_q, u_w)
}
\description{
Swap ETH for a Token, Given Token Qty check how much ETH you need
}
