% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade.R
\name{check_tok.to.eth_eth.fix}
\alias{check_tok.to.eth_eth.fix}
\title{Swap Token for ETH, Given ETH Qty check how much Token you need}
\usage{
check_tok.to.eth_eth.fix(t_a, t_d, e_q, u_w)
}
\description{
Swap Token for ETH, Given ETH Qty check how much Token you need
}
