% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade.R
\name{check_tok1.to.tok2_tok1.fix}
\alias{check_tok1.to.tok2_tok1.fix}
\title{Swap Token1 for Token2, Given Token1 Qty check how much Token2 you would get (Use Token1 -> ETH -> Token2 Route)}
\usage{
check_tok1.to.tok2_tok1.fix(t1_a, t1_d, t2_a, t2_d, t1_q, u_w)
}
\description{
Swap Token1 for Token2, Given Token1 Qty check how much Token2 you would get (Use Token1 -> ETH -> Token2 Route)
}
