% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade.R
\name{check_tok1.to.tok2_tok2.fix}
\alias{check_tok1.to.tok2_tok2.fix}
\title{Swap Token1 for Token2, Given Token2 Qty check how much Token1 you would need (Use Token1 -> ETH -> Token2 Route)}
\usage{
check_tok1.to.tok2_tok2.fix(t1_a, t1_d, t2_a, t2_d, t2_q, u_w)
}
\arguments{
\item{t1_a}{Token 1 Address}

\item{t1_d}{Token 1 Decimals}

\item{t2_a}{Token 2 Address}

\item{t2_d}{Token 2 Decimals}

\item{t2_q}{Token 2 Qty.}

\item{u_w}{Uniswap Session}
}
\value{
Numeric vector representing the Token 1 Amount Needed
}
\description{
Swap Token1 for Token2, Given Token2 Qty check how much Token1 you would need (Use Token1 -> ETH -> Token2 Route)
}
