% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade.R
\name{check_tok_balance}
\alias{check_tok_balance}
\title{Check any Token Balance}
\usage{
check_tok_balance(t_a, t_d, u_w)
}
\arguments{
\item{t_a}{Token Address}

\item{t_d}{Token Decimals}

\item{u_w}{Uniswap Session}
}
\value{
Numeric vector representing the User Token balance
}
\description{
Check any Token Balance
}
