% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{export_data}
\alias{export_data}
\title{Write out the data object given file name}
\usage{
export_data(data_to_export, path_to_export)
}
\arguments{
\item{data_to_export}{Object containing data we want to export}

\item{path_to_export}{Path of the CSV file, we want to export to}
}
\value{
Character vector indicating status of the write
}
\description{
Write out the data object given file name
}
\examples{
\dontrun{
data_to_export <- token_stats_hist_v2(token_address = "0x1f9840a85d5af5bf1d1762f925bdaddc4201f984")
path_to_export <- "~/Desktop/uniswappeR_export.csv"
export_data(data_to_export,path_to_export)
}
}
