% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{export_plot}
\alias{export_plot}
\title{Write out the analysis plots}
\usage{
export_plot(plot_to_export, path_to_export, width = 7, height = 7)
}
\arguments{
\item{plot_to_export}{Object containing plot we want to export}

\item{path_to_export}{Path of the .png file, we want to export to}

\item{width}{Width of plot in inches}

\item{height}{Height of plot in inches}
}
\value{
Character vector of the status of the write
}
\description{
Write out the analysis plots
}
\examples{
\dontrun{
plot_to_export <- vis_uniswap_stats_hist_v2()
path_to_export <- "~/Desktop/uniswappeR_plot_export.png"
export_plot(plot_to_export,path_to_export)
}
}
