% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{factory_stats}
\alias{factory_stats}
\title{Get Uniswap Factory Stats}
\usage{
factory_stats()
}
\value{
Data on the Uniswap Factory contract
}
\description{
Get Uniswap Factory Stats
}
\examples{

factory_stats()
}
