% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade.R
\name{get_infura_node}
\alias{get_infura_node}
\title{Gets the Infura Node}
\usage{
get_infura_node()
}
\value{
Your Infura Node, if set
}
\description{
Gets the Infura Node
}
\examples{
\dontrun{
get_infura_node()
}
}
