% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{pair_burn_txs_v3}
\alias{pair_burn_txs_v3}
\title{Get UniswapV3 Burn Transactions in a pair}
\usage{
pair_burn_txs_v3(pair_address = "0x1d42064fc4beb5f8aaf85f4617ae8b3b5b8bd801")
}
\arguments{
\item{pair_address}{Pair's Address}
}
\value{
Data frame representing Burn Transactions in a pair
}
\description{
Get UniswapV3 Burn Transactions in a pair
}
\examples{
\dontrun{
pair_burn_txs_v3(pair_address = "0x1d42064fc4beb5f8aaf85f4617ae8b3b5b8bd801")
}
}
