% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{pair_liq_positions_hist_v2}
\alias{pair_liq_positions_hist_v2}
\title{Get Historical Liquidity Positions in a pair}
\usage{
pair_liq_positions_hist_v2(
  pair_address = "0xd3d2e2692501a5c9ca623199d38826e513033a17"
)
}
\arguments{
\item{pair_address}{Pair's Address}
}
\value{
Historical Liquidity Positions in a pair
}
\description{
Get Historical Liquidity Positions in a pair
}
\examples{

pair_liq_positions_hist_v2(pair_address = "0xd3d2e2692501a5c9ca623199d38826e513033a17")
}
