% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{pair_liq_positions_v2}
\alias{pair_liq_positions_v2}
\title{Get Current Liquidity Positions in a pair}
\usage{
pair_liq_positions_v2(
  pair_address = "0x360b9726186c0f62cc719450685ce70280774dc8"
)
}
\arguments{
\item{pair_address}{Pair's Address}
}
\value{
Current Liquidity Positions in a pair
}
\description{
Get Current Liquidity Positions in a pair
}
\examples{

pair_liq_positions_v2(pair_address = "0x360b9726186c0f62cc719450685ce70280774dc8")
}
