% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{pair_liq_positions_v3}
\alias{pair_liq_positions_v3}
\title{Get UniswapV3 Current Liquidity Positions in a pair}
\usage{
pair_liq_positions_v3(
  pair_address = "0x1d42064fc4beb5f8aaf85f4617ae8b3b5b8bd801"
)
}
\arguments{
\item{pair_address}{Pair's Address}
}
\value{
Data frame representing Current Liquidity Positions in a pair
}
\description{
Get UniswapV3 Current Liquidity Positions in a pair
}
\examples{
\dontrun{
pair_liq_positions_v3(pair_address = "0x1d42064fc4beb5f8aaf85f4617ae8b3b5b8bd801")
}
}
