% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{pair_stats_hist_hourly_v2}
\alias{pair_stats_hist_hourly_v2}
\title{Get Hourly Pair Historical Stats}
\usage{
pair_stats_hist_hourly_v2(
  pair_address = "0x360b9726186c0f62cc719450685ce70280774dc8"
)
}
\arguments{
\item{pair_address}{Pair's Address}
}
\value{
Hourly Historical Data on a particular Pair
}
\description{
Get Hourly Pair Historical Stats
}
\examples{

pair_stats_hist_hourly_v2(pair_address = "0x360b9726186c0f62cc719450685ce70280774dc8")
}
