% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{pair_stats_hist_v2}
\alias{pair_stats_hist_v2}
\title{Get Pair Historical Stats}
\usage{
pair_stats_hist_v2(pair_address = "0xd3d2e2692501a5c9ca623199d38826e513033a17")
}
\arguments{
\item{pair_address}{Pair's Address}
}
\value{
Historical Data on a particular Pair
}
\description{
Get Pair Historical Stats
}
\examples{

pair_stats_hist_v2(pair_address = "0xd3d2e2692501a5c9ca623199d38826e513033a17")
}
