% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{pair_stats_v2}
\alias{pair_stats_v2}
\title{Get UniswapV2 Pair Stats}
\usage{
pair_stats_v2(pair_address = "0xf00e80f0de9aea0b33aa229a4014572777e422ee")
}
\arguments{
\item{pair_address}{Pair's Address}
}
\value{
List data on a particular Pair
}
\description{
Get UniswapV2 Pair Stats
}
\examples{
\dontrun{
pair_stats_v2(pair_address = "0xf00e80f0de9aea0b33aa229a4014572777e422ee")
}
}
