% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{pair_stats_v3}
\alias{pair_stats_v3}
\title{Get UniswapV3 Pair Stats}
\usage{
pair_stats_v3(pair_address = "0x1d42064fc4beb5f8aaf85f4617ae8b3b5b8bd801")
}
\arguments{
\item{pair_address}{Pair's Address}
}
\value{
List data on a particular Pair
}
\description{
Get UniswapV3 Pair Stats
}
\examples{
\dontrun{
pair_stats_v3(pair_address = "0x1d42064fc4beb5f8aaf85f4617ae8b3b5b8bd801")
}
}
