% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{pair_swap_txs_v2}
\alias{pair_swap_txs_v2}
\title{Get Swap Transactions in a pair}
\usage{
pair_swap_txs_v2(pair_address = "0x360b9726186c0f62cc719450685ce70280774dc8")
}
\arguments{
\item{pair_address}{Pair's Address}
}
\value{
Swap Transactions in a pair
}
\description{
Get Swap Transactions in a pair
}
\examples{

pair_swap_txs_v2(pair_address = "0x360b9726186c0f62cc719450685ce70280774dc8")
}
