% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{pairs_all_v2}
\alias{pairs_all_v2}
\title{Get UniswapV2 All Pairs}
\usage{
pairs_all_v2()
}
\value{
Data frame representing All Pair Data
}
\description{
Get UniswapV2 All Pairs
}
\examples{
\dontrun{
pairs_all_v2()
}
}
