% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade.R
\name{py_int}
\alias{py_int}
\title{Python builtin functions}
\usage{
py_int(x)
}
\arguments{
\item{x}{The value to pass through the python int function}
}
\value{
An R integer converted from a Python integer
}
\description{
Python builtin functions
}
