% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade.R
\name{r_num}
\alias{r_num}
\title{python object to R}
\usage{
r_num(val)
}
\arguments{
\item{val}{Numeric converter to export}
}
\value{
An R numeric vector converted from a Python numeric vector
}
\description{
python object to R
}
