% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade.R
\name{set_infura_node}
\alias{set_infura_node}
\title{Sets the Infura Node}
\usage{
set_infura_node(infura_node)
}
\arguments{
\item{infura_node}{The Infura Node}
}
\value{
A boolean TRUE if the INFURA NODE was successfully set
}
\description{
Sets the Infura Node
}
\examples{
\dontrun{
set_infura_node("https://mainnet.infura.io/v3/XXXXXXXXXXXXXXXXXXX")
}
}
