% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positions.R
\name{swap_performance}
\alias{swap_performance}
\title{Get a visualization on the swap performance}
\usage{
swap_performance(swap_data)
}
\arguments{
\item{swap_data}{The data on swaps as generated by the swaps() function}
}
\value{
ggplot2 object of visualizations on the swap performance for the given account owner
}
\description{
Get a visualization on the swap performance
}
\examples{

addresses <- c("0xb1b117a45aD71d408eb55475FC3A65454edCc94A",
"0x41D2a18E1DdACdAbFDdADB62e9AEE67c63070b76",
"0x0De20c4bDBE0d0EEFFd2956Be4c148CA86C6cC45")

swap_data <- swaps(addresses)
}
