% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positions.R
\name{swap_statistics}
\alias{swap_statistics}
\title{Get statistics on the swaps data for a given address or addresses}
\usage{
swap_statistics(swap_data, aggregate_addresses = TRUE)
}
\arguments{
\item{swap_data}{The data on swaps as generated by the swaps() function}

\item{aggregate_addresses}{If TRUE, aggregate the addresses passed in}
}
\value{
dataframe of statistics on the swaps for the given account owner
}
\description{
Get statistics on the swaps data for a given address or addresses
}
\examples{

addresses <- c("0xb1b117a45aD71d408eb55475FC3A65454edCc94A",
"0x41D2a18E1DdACdAbFDdADB62e9AEE67c63070b76",
"0x0De20c4bDBE0d0EEFFd2956Be4c148CA86C6cC45")

swap_data <- swaps(addresses)
swap_statistics(swap_data)
}
