% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{token_pair_map_v2}
\alias{token_pair_map_v2}
\title{Get UniswapV2 Token's associated pairs}
\usage{
token_pair_map_v2(token_address = "0x1f9840a85d5af5bf1d1762f925bdaddc4201f984")
}
\arguments{
\item{token_address}{Token's Address}
}
\value{
Data frame representing associated Pairs of a particular Token
}
\description{
Get UniswapV2 Token's associated pairs
}
\examples{
\dontrun{
token_pair_map_v2(token_address = "0x1f9840a85d5af5bf1d1762f925bdaddc4201f984")
}
}
