% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{token_stats_hist}
\alias{token_stats_hist}
\title{Get Token Historical Stats (Max 1000 Entries)}
\usage{
token_stats_hist(token_address = "0x1f9840a85d5af5bf1d1762f925bdaddc4201f984")
}
\arguments{
\item{token_address}{Token's Address}
}
\value{
Historical Data on a particular Token
}
\description{
Get Token Historical Stats (Max 1000 Entries)
}
\examples{

token_stats_hist(token_address = "0x1f9840a85d5af5bf1d1762f925bdaddc4201f984")
}
