% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{token_stats_hist_v3}
\alias{token_stats_hist_v3}
\title{Get UniswapV3 Token Historical Stats}
\usage{
token_stats_hist_v3(
  token_address = "0x1f9840a85d5af5bf1d1762f925bdaddc4201f984"
)
}
\arguments{
\item{token_address}{Token's Address}
}
\value{
Data frame representing historical Data on a particular Token
}
\description{
Get UniswapV3 Token Historical Stats
}
\examples{
\dontrun{
token_stats_hist_v3(token_address = "0x1f9840a85d5af5bf1d1762f925bdaddc4201f984")
}
}
