% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade.R
\name{trade_eth.to.tok_eth.fix}
\alias{trade_eth.to.tok_eth.fix}
\title{Swap ETH for a Token, Receive Tokens for specified ETH amount}
\usage{
trade_eth.to.tok_eth.fix(t_a, t_d, e_q, u_w)
}
\description{
Swap ETH for a Token, Receive Tokens for specified ETH amount
}
