% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade.R
\name{trade_eth.to.tok_tok.fix}
\alias{trade_eth.to.tok_tok.fix}
\title{Swap ETH for a Token, Buy specified fixed Token Amount}
\usage{
trade_eth.to.tok_tok.fix(t_a, t_d, t_q, u_w)
}
\description{
Swap ETH for a Token, Buy specified fixed Token Amount
}
