% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade.R
\name{trade_tok.to.eth_eth.fix}
\alias{trade_tok.to.eth_eth.fix}
\title{Swap Token for ETH, Swap tokens to receive specified ETH amount}
\usage{
trade_tok.to.eth_eth.fix(t_a, t_d, e_q, u_w)
}
\description{
Swap Token for ETH, Swap tokens to receive specified ETH amount
}
