% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade.R
\name{trade_tok1.to.tok2_tok1.fix}
\alias{trade_tok1.to.tok2_tok1.fix}
\title{Swap Token1 for Token2, Receive Token2 for specified Token1 Amount}
\usage{
trade_tok1.to.tok2_tok1.fix(t1_a, t1_d, t2_a, t2_d, t1_q, u_w)
}
\description{
Swap Token1 for Token2, Receive Token2 for specified Token1 Amount
}
