% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{uniswap_stats_hist_v2}
\alias{uniswap_stats_hist_v2}
\title{Get Uniswap Historical Stats}
\usage{
uniswap_stats_hist_v2()
}
\value{
Historical Data on the Uniswap Platform
}
\description{
Get Uniswap Historical Stats
}
\examples{

uniswap_stats_hist_v2()
}
