% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{user_hist_lps_v3}
\alias{user_hist_lps_v3}
\title{Get UniswapV3 Historical User Liquidity Positions}
\usage{
user_hist_lps_v3(user_address = "0xF1c206dd83ee2b8E6Ea675Cf827C93c58486B972")
}
\arguments{
\item{user_address}{User's Address}
}
\value{
Data frame representing Historical User Liquidity Positions
}
\description{
Get UniswapV3 Historical User Liquidity Positions
}
\examples{
\dontrun{
user_hist_lps_v3(user_address = "0xF1c206dd83ee2b8E6Ea675Cf827C93c58486B972")
}
}
