% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{user_lps_v2}
\alias{user_lps_v2}
\title{Get User Liquidity Positions}
\usage{
user_lps_v2(user_address = "0x2e3381202988d535e8185e7089f633f7c9998e83")
}
\arguments{
\item{user_address}{User's Address}
}
\value{
User User Liquidity Positions
}
\description{
Get User Liquidity Positions
}
\examples{

user_lps_v2(user_address = "0x2e3381202988d535e8185e7089f633f7c9998e83")
}
