% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{user_mints_v2}
\alias{user_mints_v2}
\title{Get User Mint Txs}
\usage{
user_mints_v2(user_address = "0x911605012f87A3017322c81fCB4C90ADA7C09116")
}
\arguments{
\item{user_address}{User's Address}
}
\value{
User Mint Txs
}
\description{
Get User Mint Txs
}
\examples{

user_mints_v2(user_address = "0x911605012f87A3017322c81fCB4C90ADA7C09116")
}
