% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{user_swaps_v3}
\alias{user_swaps_v3}
\title{Get UniswapV3 User Swap Txs}
\usage{
user_swaps_v3(user_address = "0x431B5A84aCC1297Eda88259f300262F1bc3A74f3")
}
\arguments{
\item{user_address}{User's Address}
}
\value{
Data frame representing User Swap Txs
}
\description{
Get UniswapV3 User Swap Txs
}
\examples{
\dontrun{
user_swaps_v3(user_address = "0x431B5A84aCC1297Eda88259f300262F1bc3A74f3")
}
}
