% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{vis_pair_liq_positions_v2}
\alias{vis_pair_liq_positions_v2}
\title{Visualise Liquidity Positions spread in a given pair}
\usage{
vis_pair_liq_positions_v2(
  pair_address = "0xf00e80f0de9aea0b33aa229a4014572777e422ee"
)
}
\arguments{
\item{pair_address}{Pair's Address}
}
\value{
ggplot2 Plot of Liquidity Positions spread in a given pair
}
\description{
Visualise Liquidity Positions spread in a given pair
}
\examples{
vis_pair_liq_positions_v2(pair_address = "0xf00e80f0de9aea0b33aa229a4014572777e422ee")
}
