% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{vis_pair_liq_positions_v3}
\alias{vis_pair_liq_positions_v3}
\title{Visualise Liquidity Positions spread in a given pair}
\usage{
vis_pair_liq_positions_v3(
  pair_address = "0x1d42064fc4beb5f8aaf85f4617ae8b3b5b8bd801"
)
}
\arguments{
\item{pair_address}{Pair's Address}
}
\value{
ggplot2 Plot of Liquidity Positions spread in a given pair
}
\description{
Visualise Liquidity Positions spread in a given pair
}
\examples{
\dontrun{
vis_pair_liq_positions_v3(pair_address = "0x1d42064fc4beb5f8aaf85f4617ae8b3b5b8bd801")
}
}
