% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{vis_pair_stats_hist_daily_v2}
\alias{vis_pair_stats_hist_daily_v2}
\title{Visualise various growth metrics of a given pair UniswapV2}
\usage{
vis_pair_stats_hist_daily_v2(
  pair_address = "0xf00e80f0de9aea0b33aa229a4014572777e422ee"
)
}
\arguments{
\item{pair_address}{Pair's Address}
}
\value{
ggplot2 Plot of growth metrics of a given pair
}
\description{
Visualise various growth metrics of a given pair UniswapV2
}
\examples{
vis_pair_stats_hist_daily_v2(pair_address = "0xf00e80f0de9aea0b33aa229a4014572777e422ee")
}
