% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{vis_pair_stats_hist_daily_v3}
\alias{vis_pair_stats_hist_daily_v3}
\title{Visualise various growth metrics of a given pair UniswapV3}
\usage{
vis_pair_stats_hist_daily_v3(
  pair_address = "0x1d42064fc4beb5f8aaf85f4617ae8b3b5b8bd801"
)
}
\arguments{
\item{pair_address}{Pair's Address}
}
\value{
ggplot2 Plot of growth metrics of a given pair
}
\description{
Visualise various growth metrics of a given pair UniswapV3
}
\examples{
vis_pair_stats_hist_daily_v3(pair_address = "0x1d42064fc4beb5f8aaf85f4617ae8b3b5b8bd801")
}
