% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{vis_token_pair_map_v2}
\alias{vis_token_pair_map_v2}
\title{Visualise Number of pairs the token is present UniswapV2}
\usage{
vis_token_pair_map_v2(
  token_address = "0x1f9840a85d5af5bf1d1762f925bdaddc4201f984"
)
}
\arguments{
\item{token_address}{Token's Address}
}
\value{
ggplot2 Plot of Number of pairs the token is present
}
\description{
Visualise Number of pairs the token is present UniswapV2
}
\examples{
\dontrun{vis_token_pair_map_v2(token_address = "0x1f9840a85d5af5bf1d1762f925bdaddc4201f984")}
}
