% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{vis_token_stats_hist_v3}
\alias{vis_token_stats_hist_v3}
\title{Visualise various growth metrics of a given token in UniswapV3}
\usage{
vis_token_stats_hist_v3(
  token_address = "0x1f9840a85d5af5bf1d1762f925bdaddc4201f984"
)
}
\arguments{
\item{token_address}{Token's Address}
}
\value{
ggplot2 Plot of growth metrics of a given token
}
\description{
Visualise various growth metrics of a given token in UniswapV3
}
\examples{
vis_token_stats_hist_v3(token_address = "0x1f9840a85d5af5bf1d1762f925bdaddc4201f984")
}
