% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{vis_uniswap_stats_hist_v2}
\alias{vis_uniswap_stats_hist_v2}
\title{Visualise various growth metrics of the Uniswap Platform}
\usage{
vis_uniswap_stats_hist_v2()
}
\value{
Plot of growth metrics of the Uniswap Platform
}
\description{
Visualise various growth metrics of the Uniswap Platform
}
\examples{
data_to_export <- token_stats_hist_v2(token_address = "0x1f9840a85d5af5bf1d1762f925bdaddc4201f984")
path_to_export <- "~/Desktop/uniswappeR_export.csv"
vis_uniswap_stats_hist_v2()
}
