% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{vis_uniswap_stats_hist_v3}
\alias{vis_uniswap_stats_hist_v3}
\title{Visualise various growth metrics of the UniswapV3 Platform}
\usage{
vis_uniswap_stats_hist_v3()
}
\value{
ggplot plot of growth metrics of the UniswapV3 Platform
}
\description{
Visualise various growth metrics of the UniswapV3 Platform
}
\examples{
\dontrun{vis_uniswap_stats_hist_v3()}
}
