% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade.R
\name{check_eth.to.tok_eth.fix}
\alias{check_eth.to.tok_eth.fix}
\title{Swap ETH for a Token, Given ETH Qty check how much token you would get}
\usage{
check_eth.to.tok_eth.fix(t_a, t_d, e_q, u_w)
}
\arguments{
\item{t_a}{Token Address}

\item{t_d}{Token Decimals}

\item{e_q}{Ethereum Qty.}

\item{u_w}{Uniswap Session}
}
\value{
Numeric vector representing the Token Amount you get
}
\description{
Swap ETH for a Token, Given ETH Qty check how much token you would get
}
