% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade.R
\name{check_eth.to.tok_tok.fix}
\alias{check_eth.to.tok_tok.fix}
\title{Swap ETH for a Token, Given Token Qty check how much ETH you need}
\usage{
check_eth.to.tok_tok.fix(t_a, t_d, t_q, u_w)
}
\arguments{
\item{t_a}{Token Address}

\item{t_d}{Token Decimals}

\item{t_q}{Token Qty.}

\item{u_w}{Uniswap Session}
}
\value{
Numeric vector representing the ETH needed
}
\description{
Swap ETH for a Token, Given Token Qty check how much ETH you need
}
