% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade.R
\name{check_eth_balance}
\alias{check_eth_balance}
\title{Get your ETH Balance}
\usage{
check_eth_balance(u_w)
}
\arguments{
\item{u_w}{Uniswap Session}
}
\value{
Numeric vector representing the User Ethereum balance
}
\description{
Get your ETH Balance
}
