% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade.R
\name{check_tok.to.eth_tok.fix}
\alias{check_tok.to.eth_tok.fix}
\title{Swap Token for ETH, Given Token Qty check how much ETH you would get}
\usage{
check_tok.to.eth_tok.fix(t_a, t_d, t_q, u_w)
}
\arguments{
\item{t_a}{Token Address}

\item{t_d}{Token Decimals}

\item{t_q}{Token Qty.}

\item{u_w}{Uniswap Session}
}
\value{
Numeric vector representing the ETH Amount you get
}
\description{
Swap Token for ETH, Given Token Qty check how much ETH you would get
}
