% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{factory_stats_v2}
\alias{factory_stats_v2}
\title{Get UniswapV2 Factory Stats}
\usage{
factory_stats_v2()
}
\value{
List representing data on the UniswapV2 Factory contract
}
\description{
Get UniswapV2 Factory Stats
}
\examples{
\dontrun{
factory_stats_v2()
}
}
