% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{factory_stats_v3}
\alias{factory_stats_v3}
\title{Get UniswapV3 Factory Stats}
\usage{
factory_stats_v3()
}
\value{
List representing data on the UniswapV3 Factory contract
}
\description{
Get UniswapV3 Factory Stats
}
\examples{
\dontrun{
factory_stats_v3()
}
}
