% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positions.R
\name{liquidity_range_visualization}
\alias{liquidity_range_visualization}
\title{Get a visualization liquidity range estimates}
\usage{
liquidity_range_visualization(pair_address, ...)
}
\arguments{
\item{pair_address}{The address of the pair to analyze}

\item{...}{Additional arguments passed to the liquidity_range_v3 function}
}
\value{
Visualization on the liquidity range for the given pair
}
\description{
Get a visualization liquidity range estimates
}
\examples{

}
