% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{pair_burn_txs_v2}
\alias{pair_burn_txs_v2}
\title{Get UniswapV2 Burn Transactions in a pair}
\usage{
pair_burn_txs_v2(pair_address = "0xf00e80f0de9aea0b33aa229a4014572777e422ee")
}
\arguments{
\item{pair_address}{Pair's Address}
}
\value{
Data frame representing Burn Transactions in a pair
}
\description{
Get UniswapV2 Burn Transactions in a pair
}
\examples{
\dontrun{
pair_burn_txs_v2(pair_address = "0xf00e80f0de9aea0b33aa229a4014572777e422ee")
}
}
