% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade.R
\name{trade_eth.to.tok_eth.fix}
\alias{trade_eth.to.tok_eth.fix}
\title{Swap ETH for a Token, Receive Tokens for specified ETH amount}
\usage{
trade_eth.to.tok_eth.fix(t_a, t_d, e_q, u_w)
}
\arguments{
\item{t_a}{Token Address}

\item{t_d}{Token Decimals}

\item{e_q}{Ethereum Qty.}

\item{u_w}{Uniswap Session}
}
\value{
Character vector representing the Transaction Hash
}
\description{
Swap ETH for a Token, Receive Tokens for specified ETH amount
}
