% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade.R
\name{trade_tok.to.eth_tok.fix}
\alias{trade_tok.to.eth_tok.fix}
\title{Swap Token for ETH, Receive ETH after swapping specified token amount}
\usage{
trade_tok.to.eth_tok.fix(t_a, t_d, t_q, u_w)
}
\arguments{
\item{t_a}{Token Address}

\item{t_d}{Token Decimals}

\item{t_q}{Token Qty.}

\item{u_w}{Uniswap Session}
}
\value{
Character vector representing the Transaction Hash
}
\description{
Swap Token for ETH, Receive ETH after swapping specified token amount
}
