% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade.R
\name{uniswap_session}
\alias{uniswap_session}
\title{Start the python session}
\usage{
uniswap_session(node = get_infura_node(), user_add = NULL, pvt_key = NULL)
}
\arguments{
\item{node}{The Infura Node}

\item{user_add}{User Address}

\item{pvt_key}{Private Key of the user Address}
}
\value{
The python session uniswap endpoint
}
\description{
Start the python session
}
