% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{uniswap_stats_hist_v2}
\alias{uniswap_stats_hist_v2}
\title{Get UniswapV2 Historical Stats}
\usage{
uniswap_stats_hist_v2()
}
\value{
data frame representing historical Data on the Uniswap Platform
}
\description{
Get UniswapV2 Historical Stats
}
\examples{
\dontrun{
uniswap_stats_hist_v2()
}
}
