% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{user_burns_v2}
\alias{user_burns_v2}
\title{Get User Burn Txs}
\usage{
user_burns_v2(user_address = "0x2502f65d77ca13f183850b5f9272270454094a08")
}
\arguments{
\item{user_address}{User's Address}
}
\value{
User Burn Txs
}
\description{
Get User Burn Txs
}
\examples{

user_burns_v2(user_address = "0x2502f65d77ca13f183850b5f9272270454094a08")
}
