% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{user_lps_v2}
\alias{user_lps_v2}
\title{Get User Liquidity Positions}
\usage{
user_lps_v2(user_address = "0x2502f65d77ca13f183850b5f9272270454094a08")
}
\arguments{
\item{user_address}{User's Address}
}
\value{
User Liquidity Positions
}
\description{
Get User Liquidity Positions
}
\examples{

user_lps_v2(user_address = "0x2502f65d77ca13f183850b5f9272270454094a08")
}
