% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{user_mints_v2}
\alias{user_mints_v2}
\title{Get UniswapV2 User Mint Txs}
\usage{
user_mints_v2(user_address = "0xcd8aa390e6eabbd2169b3580c1f7ce854675fd03")
}
\arguments{
\item{user_address}{User's Address}
}
\value{
Data frame representing User Mint Txs
}
\description{
Get UniswapV2 User Mint Txs
}
\examples{
\dontrun{
user_mints_v2(user_address = "0xcd8aa390e6eabbd2169b3580c1f7ce854675fd03")
}
}
