% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{user_swaps_v2}
\alias{user_swaps_v2}
\title{Get User Swap Txs}
\usage{
user_swaps_v2(user_address = "0xcd8aa390e6eabbd2169b3580c1f7ce854675fd03")
}
\arguments{
\item{user_address}{User's Address}
}
\value{
User Swap Txs
}
\description{
Get User Swap Txs
}
\examples{

user_swaps_v2(user_address = "0xcd8aa390e6eabbd2169b3580c1f7ce854675fd03")
}
