% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{vis_uniswap_stats_hist_v2}
\alias{vis_uniswap_stats_hist_v2}
\title{Visualise various growth metrics of the UniswapV2 Platform}
\usage{
vis_uniswap_stats_hist_v2()
}
\value{
ggplot2 plot of growth metrics of the UniswapV2 Platform
}
\description{
Visualise various growth metrics of the UniswapV2 Platform
}
\examples{
\dontrun{vis_uniswap_stats_hist_v2()}
}
